+++ src/common_clj/integrant_components/config.clj
@@ -9,14 +9,13 @@
   [file-path :- s/Str]
   (edn/read-string (slurp file-path)))
 
-(defmethod ig/init-key :common-clj.integrant-components/config
-  [_
-   {:keys [path env overrides]}]
-  (log/info :starting-config)
+(defmethod ig/init-key ::config
+  [_ {:keys [path env overrides]}]
+  (log/info :starting ::config)
   (merge (-> (config-file! path) env)
          overrides
          {:current-env env}))
 
-(defmethod ig/halt-key! :common-clj.integrant-components/config
+(defmethod ig/halt-key! ::config
   [_ _routes]
-  (log/info :stopping-config))
+  (log/info :stopping ::config))