+++ src/common_clj/integrant_components/routes.clj
@@ -2,12 +2,11 @@
   (:require [integrant.core :as ig]
             [taoensso.timbre :as log]))
 
-(defmethod ig/init-key :common-clj.integrant-components/routes
-  [_
-   {:keys [routes]}]
-  (log/info :starting-routes)
+(defmethod ig/init-key ::routes
+  [_ {:keys [routes]}]
+  (log/info :starting ::routes)
   (into #{} routes))
 
-(defmethod ig/halt-key! :common-clj.integrant-components/routes
+(defmethod ig/halt-key! ::routes
   [_ _routes]
-  (log/info :stopping-routes))
+  (log/info :stopping ::routes))