+++ src/common_clj/integrant_components/sqs_producer.clj
@@ -1,23 +1,49 @@
 (ns common-clj.integrant-components.sqs-producer
   (:require [amazonica.aws.sqs :as sqs]
+            [common-clj.traceability.core :as common-traceability]
             [integrant.core :as ig]
+            [medley.core :as medley]
+            [schema.core :as s]
             [taoensso.timbre :as log]))
 
-(defmethod ig/init-key :common-clj.integrant-components.sqs-producer/sqs-producer
+(defmulti produce!
+  (fn [_ {:keys [current-env]}]
+    current-env))
+
+(s/defmethod produce! :prod
+  [{:keys [queue payload]}
+   {:keys [aws-credentials]}]
+  (let [payload' (assoc payload :meta {:correlation-id (-> (common-traceability/current-correlation-id)
+                                                           common-traceability/correlation-id-appended)})]
+    (sqs/send-message aws-credentials
+                      (sqs/get-queue-url aws-credentials queue)
+                      (prn-str payload'))))
+
+(s/defmethod produce! :test
+  [{:keys [queue payload]}
+   {:keys [produced-messages]}]
+  (let [payload' (assoc payload :meta {:correlation-id (-> (common-traceability/current-correlation-id)
+                                                           common-traceability/correlation-id-appended)})]
+    (swap! produced-messages conj {:queue   queue
+                                   :payload payload'})))
+
+(defmethod ig/init-key ::sqs-producer
   [_ {:keys [components]}]
-  (log/info :starting :common-clj.integrant-components.sqs-producer/sqs-producer)
-  (let [env (-> components :config :current-env)
-        aws-credentials {:access-key (-> components :config env :aws-credentials :access-key)
-                         :secret-key (-> components :config env :aws-credentials :secret-key)
-                         :endpoint   (-> components :config env :aws-credentials :endpoint)}]
+  (log/info :starting ::sqs-producer)
+  (let [aws-credentials {:access-key (-> components :config :aws-credentials :access-key)
+                         :secret-key (-> components :config :aws-credentials :secret-key)
+                         :endpoint   (-> components :config :aws-credentials :endpoint)}]
 
     (try (sqs/list-queues aws-credentials)
          (catch Exception ex
            (log/error :invalid-credentials :exception ex)
            (throw ex)))
 
-    {:aws-credential aws-credentials}))
+    (medley/assoc-some {:current-env     (-> components :config :current-env)
+                        :aws-credentials aws-credentials}
+                       :produced-messages (when (= (-> components :config :current-env) :test)
+                                            (atom [])))))
 
-(defmethod ig/halt-key! :common-clj.integrant-components.sqs-producer/sqs-producer
+(defmethod ig/halt-key! ::sqs-producer
   [_ _sqs-producer]
-  (log/info :stopping :common-clj.integrant-components.sqs-producer/sqs-producer))
+  (log/info :stopping ::sqs-producer))